[FIX] comment user profile
이이잉 기모링~ 해결했따
